refactor frontend template application to isolate all the components for the generator to globally transform and input the setup
